requirements.txt
google-adk
google-cloud-aiplatform
pydantic
requests
pytest
